fix video in remote item channel launch player scroll playlist edit
