+++ app/build.gradle
@@ -245,9 +245,9 @@ dependencies {
     testImplementation "androidx.test.espresso:espresso-core:$ver_espresso_core"
     testImplementation "androidx.test.espresso:espresso-contrib:$ver_espresso_core"
     testImplementation "androidx.test.espresso:espresso-intents:$ver_espresso_core"
+    testImplementation "io.insert-koin:koin-test-junit4:$ver_koin"
 
     androidTestImplementation "io.insert-koin:koin-test:$ver_koin"
-    testImplementation "io.insert-koin:koin-test-junit4:$ver_koin"
     androidTestImplementation "com.google.truth:truth:$ver_truth"
     androidTestImplementation "org.mockito:mockito-core:$ver_mockito"
     androidTestImplementation "org.mockito:mockito-android:$ver_mockito"
