+++ app/src/main/java/uk/co/sentinelweb/cuer/app/util/wrapper/YoutubeJavaApiWrapper.kt
@@ -7,9 +7,8 @@ import androidx.annotation.VisibleForTesting
 import com.google.android.youtube.player.YouTubeApiServiceUtil.isYouTubeApiServiceAvailable
 import com.google.android.youtube.player.YouTubeInitializationResult.SUCCESS
 import com.google.android.youtube.player.YouTubeIntents.*
-import uk.co.sentinelweb.cuer.domain.ChannelDomain
 import uk.co.sentinelweb.cuer.domain.MediaDomain
-import uk.co.sentinelweb.cuer.domain.PlaylistDomain
+import uk.co.sentinelweb.cuer.domain.platform.YoutubeUrl.Companion.videoUrl
 
 class YoutubeJavaApiWrapper(
     private val activity: Activity,
@@ -90,20 +89,4 @@ class YoutubeJavaApiWrapper(
             .apply { activity.startActivity(this) }
     }.isSuccess
 
-    companion object {
-        fun channelUrl(media: MediaDomain) =
-            "https://youtube.com/channel/${media.channelData.let { it.customUrl ?: it.platformId }}"
-
-        fun channelUrl(channel: ChannelDomain) =
-            "https://youtube.com/channel/${channel.let { it.customUrl ?: it.platformId }}"
-
-        fun videoUrl(media: MediaDomain) = "https://www.youtube.com/watch?v=${media.platformId}"
-        fun videoUrl(platformId: String) = "https://www.youtube.com/watch?v=$platformId"
-        fun videoShortUrl(media: MediaDomain) = "https://youtu.be/${media.platformId}"
-
-        fun playlistUrl(playlist: PlaylistDomain) =
-            "https://www.youtube.com/playlist?list=${playlist.platformId}"
-
-        fun playlistUrl(platformId: String) = "https://www.youtube.com/playlist?list=${platformId}"
-    }
 }
